feat: implement ISB debug builders and enhance UI components
- Introduced new file for pipeline ISB debug builders to manage stream and KV resource metadata.
- Added functions to generate expected ISB stream and KV resources based on pipeline specifications.
- Enhanced UI components to support refresh functionality for ISB debug information.
- Updated tests to verify the new refresh behavior and ensure proper rendering of ISB debug data.
